sw_services: xilsecure: Fix implicit conversion of operands
MISRA C violation rule 10.3: Implicit conversion of operand from essential enum to signed/unsigned int.
Soln: (i)Fixed by changing type of first argument of XSecure_ProcessIpiWithPayload0() to XSecure_ProcessIpiWithPayload5()
functions from u32 to enum.
(ii)Typecasted enum to signed int.


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilsecure/src/versal/client/xsecure_aesclient.c
@@ -90,8 +90,8 @@ int XSecure_AesEncryptInit(XSecure_AesKeySource KeySrc, u32 Size, u64 IvAddr)
 	}
 
 	AesParams->IvAddr = IvAddr;
-	AesParams->OperationId = XSECURE_ENCRYPT;
-	AesParams->KeySrc = KeySrc;
+	AesParams->OperationId = (u32)XSECURE_ENCRYPT;
+	AesParams->KeySrc = (u32)KeySrc;
 	AesParams->KeySize = Size;
 	Buffer = (u64)(UINTPTR)AesParams;
 
@@ -129,8 +129,8 @@ int XSecure_AesDecryptInit(XSecure_AesKeySource KeySrc, u32 Size, u64 IvAddr)
 	}
 
 	AesParams->IvAddr = IvAddr;
-	AesParams->OperationId = XSECURE_DECRYPT;
-	AesParams->KeySrc = KeySrc;
+	AesParams->OperationId = (u32)XSECURE_DECRYPT;
+	AesParams->KeySrc = (u32)KeySrc;
 	AesParams->KeySize = Size;
 	Buffer = (u64)(UINTPTR)AesParams;
 
@@ -330,7 +330,7 @@ int XSecure_AesKeyZero(XSecure_AesKeySource KeySrc)
 	volatile int Status = XST_FAILURE;
 
 	Status = XSecure_ProcessIpiWithPayload1(XSECURE_API_AES_KEY_ZERO,
-			KeySrc);
+			(u32)KeySrc);
 
 	return Status;
 }
@@ -358,7 +358,7 @@ int XSecure_AesWriteKey(XSecure_AesKeySource KeySrc, u32 Size, u64 KeyAddr)
 	volatile int Status = XST_FAILURE;
 
 	Status = XSecure_ProcessIpiWithPayload4(XSECURE_API_AES_WRITE_KEY, Size,
-			KeySrc, (u32)KeyAddr, (u32)(KeyAddr >> 32));
+			(u32)KeySrc, (u32)KeyAddr, (u32)(KeyAddr >> 32));
 
 	return Status;
 }
